requirements.txt
openai~=1.45.1
streamlit~=1.38.0
faiss-cpu~=1.8.0.post1
sentence_transformers
langchain~=0.3.0
numpy~=1.26.4
sentence-transformers~=3.1.0
pip~=24.2
attrs~=24.2.0
distro~=1.9.0
protobuf~=5.28.1
pillow~=10.4.0
filelock~=3.16.0